requirements.txt
flask==3.0.0
boto3==1.34.0
python-dotenv==1.0.0
gunicorn==21.2.0